requirements.txt
# Core
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0

# API Framework
fastapi>=0.110
uvicorn>=0.27

# HTTP
httpx>=0.27

# Logging & CLI
loguru>=0.7
click>=8.1
rich>=13.0

# Database
sqlalchemy>=2.0

# ML / Clustering
scikit-learn>=1.4
numpy>=1.26

# Search Intelligence
pytrends>=4.9

# Web Scraping
beautifulsoup4>=4.12
requests>=2.31
lxml>=5.0

# LLM
openai>=1.12
anthropic>=0.18

# Development
pytest>=8.0
ruff>=0.3
mypy>=1.8
